vtls: split some generic sectransp routines out into a separate file
Moves routines into the new file, and cleans up some warnings in/around them.

Based on a commit by Conrad Kramer <conrad@conradkramer.com> in #17506
